requirements.txt
streamlit==1.22.0
tensorflow-cpu==2.11.0
keras==2.11.0
protobuf==3.19.6
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.2.2
requests
Pillow
h5py
